% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionLinearRegressionDiagnostics.R
\name{diagPlotLM}
\alias{diagPlotLM}
\title{Interactive Diagnostic Plots for Linear Regression Models}
\usage{
diagPlotLM(x)
}
\arguments{
\item{x}{A model of class \code{lm}.}
}
\value{
An interactive shiny app with plots of residuals and influence statistics as well as thresholds recommended by Field 2005.
}
\description{
Interactive Diagnostic Plots for Linear Regression Models
}
\examples{
\dontrun{

DiagPlotLM(lm(sales ~ adverts + airplay + attract, data = mod)
}

}
